docker/joytag-infer/requirements.base.txt
fastapi==0.110.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==2.7.4
pillow==10.4.0
numpy<2
